Better way to handle types but can still be enhanced
To export all the types of the components directly from what Cesium exports, I need to have a mapping between the exports and the exports name in lowerCamelCase while it is exported as CapitalisedCamelCase.
I created a script to generate this mapping automatically from the Cesium export. However, it is likely to work only with only one version of Cesium at a time. It couples react-cesium-fiber types with the version of Cesium...

Ideally there would be a way to map over keyof objects in typescript type definition in order to change their case, but there isn't :/

Maybe I should ship the types in another package in order to avoid coupling the 2 packages.